Temp include TexSyn.h to share API and Utilities.h
Temp implementation marked in code with UTILITIES_NAMESPACES

commented out the (tiny) contents of LazyPredator/Utilities.h, replaced it with #include "../TexSyn/TexSyn.h", and that seemed to work fine. (The problem before had been the definition of Timer in LazyPredator which duplicated the one in TexSyn)

Prompted to do that so I could add a RandomSequence in each FunctionSet.