Upgrade to use "gnu++17" instead of "gnu++14" and related changes.
Changed the Xcode project setting CLANG_CXX_LANGUAGE_STANDARD to  "gnu++17" from "gnu++14".
Mostly because I wanted to use the auto destructuring binds, as in:  for(auto& [k, v] : std::map...) {...}
Rewrote the 6? of those in the code that I found.
But several of those were in FunctionSet::printSet(), which had some code duplication, so I fixed that, and renamed it FunctionSet::print() to match the methods in GpType and GpFunction.