Add TournamentGroup::getValid() and setValid().
Intended to allow (say) a TournamentFunction to decided that the current tournament is invalid (a bust, aborted, inconclusive, failed). Population::evolutionStep() will skip making any changes to the population is this flag is set false. Intended for the case in TexSyn/camouflage where the predator does not find ANY of the three prey, so they ALL should survive.